requirements.txt
python-dotenv>=1.0.0
flask==3.0.0
flask-cors==4.0.0
pandas>=2.1.0
gunicorn==21.2.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
APScheduler>=3.10.0
